✨ feat: 기본 표현함 생성 및 관련 예외 처리 추가

+++ src/main/java/com/mallang/mallang_backend/domain/sentence/expressionbook/service/impl/ExpressionBookServiceImpl.java
@@ -2,12 +2,12 @@ package com.mallang.mallang_backend.domain.sentence.expressionbook.service.impl;
 
 import com.mallang.mallang_backend.domain.member.entity.Member;
 import com.mallang.mallang_backend.domain.member.repository.MemberRepository;
+import com.mallang.mallang_backend.domain.sentence.expression.entity.Expression;
 import com.mallang.mallang_backend.domain.sentence.expression.repository.ExpressionRepository;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookRequest;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionBookResponse;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.ExpressionResponse;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.dto.savedExpressionsRequest;
-import com.mallang.mallang_backend.domain.sentence.expression.entity.Expression;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.entity.ExpressionBook;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.repository.ExpressionBookRepository;
 import com.mallang.mallang_backend.domain.sentence.expressionbook.service.ExpressionBookService;
@@ -19,13 +19,15 @@ import com.mallang.mallang_backend.domain.video.video.repository.VideoRepository
 import com.mallang.mallang_backend.global.exception.ErrorCode;
 import com.mallang.mallang_backend.global.exception.ServiceException;
 import com.mallang.mallang_backend.global.gpt.service.GptService;
-
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.time.LocalTime;
 import java.util.List;
+
+import static com.mallang.mallang_backend.global.constants.AppConstants.DEFAULT_EXPRESSION_BOOK_NAME;
+import static com.mallang.mallang_backend.global.exception.ErrorCode.EXPRESSIONBOOK_DELETE_DEFAULT_FORBIDDEN;
 
 @Service
 @RequiredArgsConstructor
@@ -47,6 +49,10 @@ public class ExpressionBookServiceImpl implements ExpressionBookService {
 
         if (!member.getSubscription().isStandardOrHigher()) {
             throw new ServiceException(ErrorCode.NO_EXPRESSIONBOOK_CREATE_PERMISSION);
+        }
+
+        if (request.getName().equals(DEFAULT_EXPRESSION_BOOK_NAME)) {
+            throw new ServiceException(ErrorCode.EXPRESSIONBOOK_CREATE_DEFAULT_FORBIDDEN);
         }
 
         ExpressionBook expressionBook = ExpressionBook.builder()
@@ -94,6 +100,10 @@ public class ExpressionBookServiceImpl implements ExpressionBookService {
 
         if (!book.getMember().getId().equals(memberId)) {
             throw new ServiceException(ErrorCode.FORBIDDEN_EXPRESSION_BOOK);
+        }
+
+        if (DEFAULT_EXPRESSION_BOOK_NAME.equals(book.getName())) {
+            throw new ServiceException(EXPRESSIONBOOK_DELETE_DEFAULT_FORBIDDEN);
         }
 
         expressionBookRepository.delete(book);
